retries and dockerfile

+++ src/main/java/com/kvdbcs/application/service/CheckInstanceHealthService.java
@@ -12,6 +12,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
 
 @Singleton
 public class CheckInstanceHealthService {
@@ -19,6 +21,7 @@ public class CheckInstanceHealthService {
     private final DbInstanceRepository repository;
     private final DbInstanceClient dbInstanceClient;
     private final ApplicationEventPublisher eventPublisher;
+    private final Map<Long, Integer> retries = new HashMap<>();
 
     public CheckInstanceHealthService(DbInstanceRepository repository, DbInstanceClient dbInstanceClient, ApplicationEventPublisher eventPublisher) {
         this.dbInstanceClient = dbInstanceClient;
@@ -26,12 +29,14 @@ public class CheckInstanceHealthService {
         this.eventPublisher = eventPublisher;
     }
 
-    @Scheduled(initialDelay = "15s", fixedDelay = "30s")
+    @Scheduled(initialDelay = "15s", fixedDelay = "10s")
     public void execute() throws IOException {
         logger.info("Starting health check for all instances");
         for (DbInstance dbInstance : repository.findAll()) {
             if (dbInstanceClient.isHealthy(dbInstance)) { continue; }
-
+            retries.put(dbInstance.getId(), retries.getOrDefault(dbInstance.getId(), 0) + 1);
+            if (retries.get(dbInstance.getId()) < 3) { continue; }
+            retries.replace(dbInstance.getId(), null);
             repository.delete(dbInstance);
             var file = new InstancesRecoveryFile(InstancesRecoveryFile.DEFAULT_PATH);
             file.deleteInstance(dbInstance);
